Minor update
The library has been updated with several minor bug fixes and example
code updates.

To provide a single point of access for information about IAP Badger
that is up to date, documentation and sample tutorials for IAP Badger
will now be maintained our website.

The standard lua version of the library for IAP Badger will continue to
be maintained on our Github repository.